clean up imports of web-local into web-common
