+++ web-common/src/components/column-profile/ColumnProfile.svelte
@@ -4,7 +4,6 @@
     createQueryServiceTableColumns,
     createQueryServiceTableRows,
   } from "@rilldata/web-common/runtime-client";
-  import { NATIVE_SELECT } from "@rilldata/web-local/lib/util/component-classes";
   import { onMount } from "svelte";
   import { runtime } from "../../runtime-client/runtime-store";
   import { getColumnType } from "./column-types";
@@ -60,6 +59,10 @@
   } else {
     sortedProfile = profile;
   }
+
+  /** classes for the native select element */
+  export const native_select =
+    "rounded border border-6 border-transparent hover:bg-gray-200";
 </script>
 
 <!-- pl-16 -->
@@ -70,7 +73,7 @@
     : '4'} pr-5 pb-2 flex justify-between text-gray-500 pt-1"
   class:flex-col={containerWidth < 325}
 >
-  <select bind:value={sortMethod} class={NATIVE_SELECT} style:font-size="11px">
+  <select bind:value={sortMethod} class={native_select} style:font-size="11px">
     <option value={sortByOriginalOrder}>show original order</option>
     <option value={defaultSort}>sort by type</option>
     <option value={sortByNullity}>sort by null %</option>
@@ -78,7 +81,7 @@
   </select>
   <select
     bind:value={mode}
-    class={NATIVE_SELECT}
+    class={native_select}
     class:hidden={containerWidth < 325}
     style:font-size="11px"
     style:transform="translateX(4px)"
