+++ web-common/src/components/virtualized-table/sections/ColumnHeaders.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import type { VirtualizedTableColumns } from "@rilldata/web-local/lib/types";
   import { createEventDispatcher } from "svelte";
   import ColumnHeader from "../core/ColumnHeader.svelte";
+  import type { VirtualizedTableColumns } from "../types";
 
   const dispatch = createEventDispatcher();
