+++ web-common/src/layout/navigation/NavigationEntry.svelte
@@ -8,7 +8,7 @@
   import { notifications } from "@rilldata/web-common/components/notifications";
   import Tooltip from "@rilldata/web-common/components/tooltip/Tooltip.svelte";
   import TooltipContent from "@rilldata/web-common/components/tooltip/TooltipContent.svelte";
-  import { createCommandClickAction } from "@rilldata/web-local/lib/util/command-click-action";
+  import { createCommandClickAction } from "../../lib/actions/command-click-action";
   import { createShiftClickAction } from "../../lib/actions/shift-click-action";
   import { emitNavigationTelemetry } from "./navigation-utils";
   import { currentHref } from "./stores";
