+++ web-common/src/lib/actions/command-click-action.ts
