--- web-local/src/lib/util/component-classes.ts
@@ -1,7 +1,0 @@
-/** classes for the native select element */
-export const NATIVE_SELECT =
-  "rounded border border-6 border-transparent hover:bg-gray-200";
-
-/** classes for elements that trigger the highlight in a model query */
-export const QUERY_REFERENCE_TRIGGER =
-  "hover:bg-yellow-200 hover:cursor-pointer";
