--- web-local/src/lib/util/os-detection.ts
@@ -1,3 +1,0 @@
-export function isMac() {
-  return window.navigator.userAgent.includes("Macintosh");
-}
